+++ CHANGELOG.rst
@@ -15,6 +15,8 @@ Fixed
 -----
 
 * Fix conducting of cycle with a fork. Fixes #169 (bug fix)
+* Fix request_workflow_status to ignore certain status change errors such as pausing a workflow
+  that is already pausing and canceling a workflow that is already canceling. (bug fix)
 
 1.0.0
 -----